pyproject.toml
[project]
name = "autotrade"
version = "0.1.0"
description = "GAT-based force-directed coin graph trading with HRM reasoning"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "coinbase-advanced-py>=1.3.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "torch>=2.0.0",
    "duckdb>=0.9.0",
    "python-dotenv>=1.0.0",
    "tiktoken>=0.5.0",
    "rustbpe>=0.1.0",
    "pyarrow>=14.0.0",
    "requests>=2.31.0",
    "mlx>=0.16.0",
]

[project.scripts]
autotrade = "autotrade.graph_showdown:main"
autotrade-ingest = "autotrade.ingest:main"
autotrade-cross = "autotrade.fetch_cross_pairs_1m:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
